misc/HtmlTestSuite/results/test-0665.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Tests correct handling of min-height: min-content with dynamic changes</title>
<link href="https://www.w3.org/TR/css-flexbox-1/#resolve-flexible-lengths" rel="help" title="4.5. Implied Minimum Size of Flex Items"/>
<link href="http://www.google.com/" rel="author" title="Google Inc."/>
<link href="support/flexbox.css" rel="stylesheet"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>
<style>
.container {
  height: 300px;
  outline: 2px solid black;
}

.inner
{
  width: 400px;
  flex: 1;
  background-color: green;
}
#container2 .flexbox &gt; * { flex-basis: 0; }
#container2 .column &gt; * { flex-basis: auto; }
.container .flexbox { min-height: min-content; }
.container &gt; .flexbox { min-height: 0; }
</style>
<script>
function change() {
  var container = document.getElementById('container');
  container.offsetHeight;
  container.style.height = '80px';
  container = document.getElementById('container2');
  container.offsetHeight;
  container.style.height = '80px';
  checkLayout('.container');
}
</script>
</head><body onload="change()">
<p>Green rectangle should be entirely within the black rectangle</p>
<div id="log"/>
<div class="container" id="container">
  <div class="flexbox column" style="height: 100%;">
    <div class="flexbox flex-one">
        <div class="flexbox column">
          <div class="flexbox column flex-one">
            <div class="inner" data-expected-height="80">
            </div>
          </div>
        </div>
    </div>
  </div>
</div>

<div class="container" id="container2">
  <div class="flexbox column" style="height: 100%;">
    <div class="flexbox flex-one">
        <div class="flexbox column">
          <div class="flexbox column flex-one">
            <div class="inner" data-expected-height="80">
            </div>
          </div>
        </div>
    </div>
  </div>
</div>
</body></html>
